requirements.txt
docker == 7.0.0
core @ git+https://github.com/alkanor/tmppubliccore.git#egg=core
paramiko == 3.4.0
cryptography == 42.0.7
psutil == 5.9.8
networkx == 3.3
pydantic == 2.7.1
pyyaml == 6.0.1
regex == 2024.5.10
z3-solver == 4.13.0.0